#35: Id dependency was moved from DAL to Core

+++ src/DAL/Implementations/EfReadWriteRepository.cs
@@ -2,17 +2,24 @@
 {
     #region << Using >>
 
+    using System;
     using System.Collections.Generic;
     using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
-    using CRUD.Extensions;
-    using Microsoft.EntityFrameworkCore;
 
     #endregion
 
-    public class EfReadWriteRepository<TEntity> : EfReadRepository<TEntity>, IReadWriteRepository<TEntity> where TEntity : EntityBase, new()
+    public class EfReadWriteRepository<TEntity> : EfReadRepository<TEntity>, IReadWriteRepository<TEntity> where TEntity : class, new()
     {
+        #region Constants
+
+        private const string EntityCanTBeNull = "Entity can't be null!";
+
+        private const string EntitiesCanTBeEmptyOrNull = "Entities can't be empty or null!";
+
+        #endregion
+
         #region Constructors
 
         public EfReadWriteRepository(IEfDbContext context) : base(context) { }
@@ -21,39 +28,63 @@
 
         #region Interface Implementations
 
-        public async Task AddOrUpdateAsync(TEntity entity, CancellationToken cancellationToken = default)
+        public async Task AddAsync(TEntity entity, CancellationToken cancellationToken = default)
         {
+            if (entity == null)
+                throw new ArgumentException(EntityCanTBeNull);
+
             var dbSet = this._context.Set<TEntity>();
 
-            if (await dbSet.AnyAsync(predicate: r => Equals(r.Id, entity.Id), cancellationToken: cancellationToken))
-                dbSet.Update(entity);
-            else
-                dbSet.Add(entity);
+            dbSet.Add(entity);
 
             await this._context.SaveChangesAsync(cancellationToken);
         }
 
-        public async Task AddOrUpdateAsync(IEnumerable<TEntity> entities, CancellationToken cancellationToken = default)
+        public async Task AddAsync(IEnumerable<TEntity> entities, CancellationToken cancellationToken = default)
         {
             var entitiesArray = entities.ToArrayOrEmpty();
 
-            if (!entitiesArray.Any())
-                return;
+            if (!entitiesArray.Any() || entitiesArray.Any(r => r == null))
+                throw new ArgumentException(EntitiesCanTBeEmptyOrNull);
 
             var dbSet = this._context.Set<TEntity>();
 
-            var entitiesIds = entitiesArray.GetIds();
-            var existingEntitiesIds = await dbSet.Where(r => entitiesIds.Contains(r.Id))
-                                                 .Select(r => r.Id).ToArrayAsync(cancellationToken);
+            await dbSet.AddRangeAsync(entitiesArray, cancellationToken);
 
-            dbSet.UpdateRange(entitiesArray.Where(r => existingEntitiesIds.Contains(r.Id)));
-            await dbSet.AddRangeAsync(entitiesArray.Where(r => !existingEntitiesIds.Contains(r.Id)), cancellationToken);
+            await this._context.SaveChangesAsync(cancellationToken);
+        }
+
+        public async Task UpdateAsync(TEntity entity, CancellationToken cancellationToken = default)
+        {
+            if (entity == null)
+                throw new ArgumentException(EntityCanTBeNull);
+
+            var dbSet = this._context.Set<TEntity>();
+
+            dbSet.Update(entity);
+
+            await this._context.SaveChangesAsync(cancellationToken);
+        }
+
+        public async Task UpdateAsync(IEnumerable<TEntity> entities, CancellationToken cancellationToken = default)
+        {
+            var entitiesArray = entities.ToArrayOrEmpty();
+
+            if (!entitiesArray.Any() || entitiesArray.Any(r => r == null))
+                throw new ArgumentException(EntitiesCanTBeEmptyOrNull);
+
+            var dbSet = this._context.Set<TEntity>();
+
+            dbSet.UpdateRange(entitiesArray);
 
             await this._context.SaveChangesAsync(cancellationToken);
         }
 
         public async Task DeleteAsync(TEntity entity, CancellationToken cancellationToken = default)
         {
+            if (entity == null)
+                throw new ArgumentException(EntityCanTBeNull);
+
             this._context.Set<TEntity>().Remove(entity);
 
             await this._context.SaveChangesAsync(cancellationToken);
@@ -62,8 +93,9 @@
         public async Task DeleteAsync(IEnumerable<TEntity> entities, CancellationToken cancellationToken = default)
         {
             var entitiesArray = entities.ToArrayOrEmpty();
-            if (!entitiesArray.Any())
-                return;
+
+            if (!entitiesArray.Any() || entitiesArray.Any(r => r == null))
+                throw new ArgumentException(EntitiesCanTBeEmptyOrNull);
 
             this._context.Set<TEntity>().RemoveRange(entitiesArray);
 
@@ -71,33 +103,5 @@
         }
 
         #endregion
-
-        public async Task DeleteAsync(object id, CancellationToken cancellationToken = default)
-        {
-            var entity = await GetByIdOrDefaultAsync(id: id, cancellationToken: cancellationToken);
-
-            this._context.Set<TEntity>().Remove(entity);
-
-            await this._context.SaveChangesAsync(cancellationToken);
-        }
-
-        public async Task DeleteAsync(IEnumerable<object> ids, CancellationToken cancellationToken = default)
-        {
-            var idsArray = ids.ToArrayOrEmpty();
-
-            if (!idsArray.Any())
-                return;
-
-            var dbSet = this._context.Set<TEntity>();
-
-            var entities = await dbSet.Where(r => idsArray.Contains(r.Id)).ToArrayAsync(cancellationToken);
-
-            if (!entities.Any())
-                return;
-
-            dbSet.RemoveRange(entities);
-
-            await this._context.SaveChangesAsync(cancellationToken);
-        }
     }
 }
